+++ src/core/mbc/nombc.cpp
@@ -2,14 +2,14 @@
 
 namespace natsukashii::core
 {
-NoMBC::NoMBC(std::vector<byte>& rom) : rom(rom) {}
+NoMBC::NoMBC(std::vector<u8>& rom) : rom(rom) {}
 
-byte NoMBC::Read(half addr)
+u8 NoMBC::Read(u16 addr)
 {
 if(addr >= 0 && addr < 0x8000)
     return rom[addr];
 return 0xff;
 }
 
-void NoMBC::Write(half addr, byte val) {}
+void NoMBC::Write(u16 addr, u8 val) {}
 } // natsukashii::core
